backend/requirements/base.txt
asgiref==3.6.0
django==4.1.7
django-cors-headers==3.14.0
django_debug_toolbar==3.8.1
django-environ==0.10.0
django-extensions==3.2.1
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.2
sqlparse==0.4.3